Follow standard Python conventions in tests
* Rename test scripts to test_*.py. This makes them discoverable by
  standard tools like IDEs, pytest, TestLoader.discover, etc.

* Automatically discover test classes instead of listing them manually.
  This is simpler and less error prone. In addition, standard command
  line options now work, for example, -k for filtering tests.

* Don't use deprecated methods and implement test_ methods instead of
  runTest.
