+++ Python/test/test_bonds.py
@@ -356,9 +356,5 @@ class AmortizingFixedRateBondTest(unittest.TestCase):
 
 
 if __name__ == "__main__":
-    print("testing QuantLib " + ql.__version__)
-    suite = unittest.TestSuite()
-    suite.addTest(unittest.makeSuite(FixedRateBondTest, "test"))
-    suite.addTest(unittest.makeSuite(FixedRateBondKwargsTest, "test"))
-    suite.addTest(unittest.makeSuite(AmortizingFixedRateBondTest, "test"))
-    unittest.TextTestRunner(verbosity=2).run(suite)
+    print("testing QuantLib", ql.__version__)
+    unittest.main(verbosity=2)
