+++ Python/test/test_coupons.py
@@ -467,11 +467,5 @@ class SubPeriodsCouponTest(unittest.TestCase):
 
 
 if __name__ == '__main__':
-    print('testing QuantLib ' + ql.__version__)
-    suite = unittest.TestSuite()
-    suite.addTest(unittest.makeSuite(CashFlowsTest, 'test'))
-    suite.addTest(unittest.makeSuite(SubPeriodsCouponTest, 'test'))
-    suite.addTest(unittest.makeSuite(IborCouponTest, 'test'))
-    suite.addTest(unittest.makeSuite(OvernightCouponTest, 'test'))
-    suite.addTest(unittest.makeSuite(FixedRateCouponTest, 'test'))
-    unittest.TextTestRunner(verbosity=2).run(suite)
+    print("testing QuantLib", ql.__version__)
+    unittest.main(verbosity=2)
