+++ Python/test/test_currencies.py
@@ -42,7 +42,5 @@ class CurrencyTest(unittest.TestCase):
 
 
 if __name__ == '__main__':
-    print('testing QuantLib ' + ql.__version__)
-    suite = unittest.TestSuite()
-    suite.addTest(unittest.makeSuite(CurrencyTest, 'test'))
-    unittest.TextTestRunner(verbosity=2).run(suite)
+    print("testing QuantLib", ql.__version__)
+    unittest.main(verbosity=2)
