+++ Python/test/test_swap.py
@@ -244,8 +244,5 @@ class EquityTotalReturnSwapTest(unittest.TestCase):
 
 
 if __name__ == '__main__':
-    print('testing QuantLib ' + ql.__version__)
-    suite = unittest.TestSuite()
-    suite.addTest(unittest.makeSuite(ZeroCouponSwapTest, 'test'))
-    suite.addTest(unittest.makeSuite(EquityTotalReturnSwapTest, 'test'))
-    unittest.TextTestRunner(verbosity=2).run(suite)
+    print("testing QuantLib", ql.__version__)
+    unittest.main(verbosity=2)
